2.4.7-3

+++ index.js
@@ -39,6 +39,7 @@ bot.on('message', async message => {
     let sender = message.author;
     let nick = sender.username;
     let Staff = message.guild.roles.find(x => x.name === "Guide");
+    let logchannel = message.guild.channels.find(x => x.name === "logs");
     if (bot.user.id === sender.id) {return};
 
 
@@ -259,9 +260,6 @@ bot.on('message', async message => {
         let args = msg.split(" ").slice(1);
         let rUser = message.guild.member(message.mentions.users.first() || message.guild.members.get(args[0]));
         let rreason = args.join(" ").slice(22);
-        let logchannel = message.guild.channels.find(x => x.name === "logs");
-
-          if(!rreason) return message.reply("You can't send an empty paper.");
 
           let mailEmbed = new Discord.RichEmbed()
           .setDescription("Today's Mail")
@@ -270,6 +268,37 @@ bot.on('message', async message => {
           .addField("Recipient:", rUser)
           .addField("Mail", rreason)
           .addField("Sent At", message.createdAt)
+
+          let mailEmbedA = new Discord.RichEmbed()
+          .setDescription("Today's Mail")
+          .setColor(0xe0782b)
+          .addField("Sent by:", `${sender} with ID: ${sender.id}`)
+          .addField("Mail", args[0])
+          .addField("Sent At", message.createdAt)
+
+          if(!rUser) {
+            message.delete()
+            logchannel.send(mailEmbedA)
+
+            message.guild.members.get(sender.id)
+          .createDM()
+                .then(dm => {
+                  dm.send({embed: {
+                    color: 0x15f153,
+                    title: "Thank you for the mail!" ,
+                  description: `Your mail to ${rUser} was sent!`,
+                  timestamp: new Date(),
+                    footer: {
+                    icon_url: "186487324517859328".avatarURL,
+                    text: "Any abuse of the mailing system will result to a penalty"
+                    }
+                  }})
+                })
+
+            return;
+          }
+
+          if(!rreason) return message.reply("You can't send an empty paper.");
           
           message.delete()
           logchannel.send(mailEmbed)
